Improved argument and options parsing
Before this enhancement, the arguments defined when scheduling a new command couldn't include white spaces, even when using quotes. This was because the way the function getArguments(true) works. Now it uses regex to do the parsing so it can handle arguments and options using double quotes, single quotes, nested quotes and no quotes.